Initial commit to develop
Interim commit for the vis repo.  Still some issues, but the project should run with 'npm start' on port 8080

Uncertain if one can checkout the repo from npm and use Chart / Histogram, but this is a start.
Likely need to create an examples directory and move ./src/pages into ./examples.

Do we want this example project to use react-router?  It's a nice example, but maybe unnecessary